==> _scripts/04b_Encounterability.Rmd <==
---
title: "encounterability"
author: "Leslie Roberson"
date: "30/12/2021"
output: html_document
---

## Summary

1) make dataframe of species-gears overlaps

2) Calculate encounterability probability (A & B probability)

    (overlap / full depth range of gear)  x  (overlap / full depth range of species)
    
    where, overlap = number of meters both species and gears are true for 

3) scale values 1-3

# Encounterability depth assumptions

We conservatively assumed all gears are deployed from the surface to 20m for drift gillnets (Aranda 2017), 280m for purse seines (Romanov 2002), and 400m for longlines (Song et al. 2009). 


```{r setup, include=FALSE}

library(here)
library(readxl)
library(tidyverse)
library(data.table)
library(pillar)

knitr::opts_chunk$set(echo = TRUE)
```

## get depth info

```{r depths}
## Load spp metadata sheet 
spp_info <- read_excel(here("_data/spp_metadata.xlsx"))

```

## Calculate encounterability overlap

```{r encounterability overlap}
#<><>< make dt of species depths
species <- spp_info %>% 
  dplyr::select(c(species_id, depth_m_min_adj, depth_m_max_adj)) %>% ## take corrected depth cols here
  distinct() %>%
  data.table()
species$depth_m_max_adj <- as.numeric(species$depth_m_max_adj)
species$depth_m_min_adj <- as.numeric(species$depth_m_min_adj)
  
species %>% filter(is.na(depth_m_min_adj)) # no missing depths 

##<><>< make dt of gear depths
gears <- data.table(
  gear = c('GND', 'PST', 'LLT'),
  gear_min = c(0, 0, 0),
  gear_max = c(20, 280, 400) 
)

##<><><> overlap
require(data.table)
setkey(gears, gear_min, gear_max)
overlaps <- foverlaps(species, gears,
                      by.x = c('depth_m_min_adj', 'depth_m_max_adj'),
                      by.y = c('gear_min', 'gear_max')
                      )
overlaps[,ol_length := pmin(depth_m_max_adj, gear_max)-pmax(depth_m_min_adj, gear_min)]
View(overlaps)
overlaps_df <- data.frame(overlaps)

## make species encounterability df
nrow(overlaps_df) # n = 962
n_distinct(overlaps_df$species_id) # n = 405
spp_encount <- spp_info %>%
  dplyr::select(c(species_id, class, common_name, sciname, class, grp_code, depth_m_min_adj, depth_m_max_adj)) %>%
  left_join(overlaps_df, by = c("species_id", "depth_m_min_adj", "depth_m_max_adj"))

## Calculate A&B probability of encounter
glimpse(spp_encount)
spp_encount <- spp_encount %>%
  mutate(gear_range = (gear_max - gear_min),
         spp_range = (depth_m_max_adj - depth_m_min_adj),
         encount_prob = (ol_length/gear_range)*(ol_length/spp_range)) 
spp_encount %>% dplyr::select(species_id, depth_m_min_adj:encount_prob)
spp_encount %>% filter(is.na(encount_prob)) %>% distinct(species_id) # 26 species don't overlap with any gear

## save encounterability df
write_csv(spp_encount, here("_results/Susceptibility/Encounterability/spp_encount_long.csv"))

## remove zeros
encount_sub <- spp_encount %>% dplyr::select(c(species_id, gear, encount_prob))
head(encount_sub)
unique(encount_sub$gear) 
n_distinct(encount_sub$species_id) # 405
encount_sub <- encount_sub %>% filter(!is.na(gear))
nrow(encount_sub) 
encount_sub %>% arrange(encount_prob)
## remove E==0 species
encount_sub <- encount_sub %>% filter(encount_prob>0)

```

Scale 1-3

For Encounterability, scale individually for each gear because those dynamics are gear-specific
- so a high horizontal overlap for gillnets is not the same as for longlines, becuase gillnet effort will be much more concentrated within that depth range compared to longlines, which could be set at different depths within that range 

```{scale encount}

## scale encounterability per gear 1-3
encount_sub <- encount_sub %>%
  group_by(gear) %>%
  mutate(E_scaled = scales::rescale(encount_prob, to= c(1,3)))
encount_sub %>% arrange(desc(E_scaled))

## fix species names to match availability rasters
encount_sub$species_id <- stringr::str_replace_all(encount_sub$species_id, "-", "_")

## save
write_csv(encount_sub, here("_results/Susceptibility/Encounterability/sum_encount_allgears.csv"))

```
