==> _scripts/04d_Suscept123_compare.Rmd <==
---
title: "susceptibility sensitivity analysis scenario"
author: "Leslie Roberson"
date: "02/01/2022"
output: html_document
---

## Summary

Here I compare the results of our adapted PSA approach (using continuous axes) to the traditional 1,2,3 score based method

```{r setup, include=FALSE}

library(pillar)
library(here)
library(data.table)
library(tidyverse)
library(rlist) # install.packages("rlist")
library(raster)
library(purrr)
library(sf)
library(sp)

knitr::opts_chunk$set(echo = TRUE)
```

# Score availability

Use the Proportion of range exposed to fishing
use equal width bins:

1: 0-33.3% overlap
2: 33.4-66.6 % overlap
3. 66.7-100% overlap

```{r score availability}

## load species ranges and availability results
avail <- read_csv(here("_results/Susceptibility/Availability/sum_availcells_allgears.csv"))
glimpse(avail)
ranges <- read_csv(here("_data/species_ranges.csv"))
glimpse(ranges)

avail123 <- left_join(avail, ranges, by="species_id")
## calculate exposure
avail123 <- avail123 %>% 
  mutate(prop_exposure=ncells_overlap/ncells_spp)

avail123 <- avail123 %>% 
  group_by(species_id, GEAR) %>%
  mutate(avail123=case_when(
    is.na(prop_exposure)~as.numeric(0), # just in case
    prop_exposure<0.334~as.numeric(1),
    prop_exposure>0.667~as.numeric(3),
    TRUE~as.numeric(2)
  ))
avail123 %>% arrange(prop_exposure) 
summary(avail123$prop_exposure)
# Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
# 0.5000  0.7747  0.9015  0.8658  0.9476  1.0000 

## save intermed file
write_csv(avail123, here("_results/Susceptibility/Availability/avail123_species_allgears.csv"))

```

## score encounterability

```{r 123 encount}

## save encounterability df
encount <- read_csv(here("_results/Susceptibility/Encounterability/spp_encount_long.csv"))
glimpse(encount)

## calculate % overlap
encount0123 <- encount %>%
  group_by(species_id, gear) %>%
  mutate(overlap_prop=ol_length/spp_range,
         encount123=case_when(
           is.na(overlap_prop)~as.numeric(0), # NA is when there's no overlap with that gear
           overlap_prop<0.25~as.numeric(1),
           overlap_prop>0.5~as.numeric(3),
           TRUE~as.numeric(2)
         ))
summary(encount0123)

## remove species/gear interactions with 0 encouterability
encount0123 <- encount0123 %>% filter(encount_prob>0)
glimpse(encount0123)

## save intermediary file
write_csv(encount0123, here("_results/Susceptibility/Encounterability/encount123_species_allgears.csv"))

```

## Get selectivity and PCM scores

(Allowing zero selectivity interactions to be removed)

```{r selectivity PCM scores}

## results from spatial PSA (same Selectivity scores)
S.scores <- read_csv(here("_results/Susceptibility/S_scores.csv"))
glimpse(S.scores)

sel0123 <- S.scores %>%
  dplyr::select(c(species_id, gear, sel0123))
glimpse(sel0123) # 730 non-zero interactions

pcm123 <- S.scores %>%
  dplyr::select(c(species_id, gear, pcm))

```

Calculate geometric mean

```{r geo mean}

## combine scores
glimpse(avail123)
glimpse(encount0123)
glimpse(sel0123)
glimpse(pcm123)

## change availability species ids to match (had _ instead of - as arcGIS artefact)
avail123 <- avail123 %>% 
  mutate(species_id=stringr::str_replace_all(species_id, "_", "-")) %>%
  rename(gear=GEAR)

AE <- encount0123 %>% 
  left_join(avail123, by=c("species_id", "gear")) 
glimpse(suscept0123)# 921 non-zero overlaps

selpcm <- left_join(sel0123, pcm123, by=c("species_id", "gear")) 
glimpse(selpcm)

suscept0123 <- inner_join(AE, selpcm, by=c("species_id", "gear"))
glimpse(suscept0123)# 730 non zero interactions

## calculate geo mean
suscept0123$suscept0123 <- apply(suscept0123[ , c('avail123', 'encount123', 'sel0123', 'pcm')], 1, function(x) exp(mean(log(x)))) 
# 1 means apply by row
suscept0123 %>% filter(suscept0123==0) # none
## save
write_csv(suscept0123, here("_results/Susceptibility/S_0123_scores.csv"))

```
